chore(sdk): generate go extend sdk (230409.2048)
- generated from source sdk spec commit: 9de2d33c81b4c247b92c282b8990e33887a08231

+++ dsmc-sdk/pkg/dsmcclientmodels/models_create_image_patch_request.go
@@ -31,7 +31,7 @@ type ModelsCreateImagePatchRequest struct {
 
 	// image size
 	// Required: true
-	ImageSize *int32 `json:"imageSize"`
+	ImageSize *int64 `json:"imageSize"`
 
 	// namespace
 	// Required: true
