+++ dsmc-sdk/pkg/dsmcclientmodels/models_image_record.go
@@ -36,7 +36,7 @@ type ModelsImageRecord struct {
 
 	// image size
 	// Required: true
-	ImageSize *int32 `json:"imageSize"`
+	ImageSize *int64 `json:"imageSize"`
 
 	// modified by
 	// Required: true
